Move Filer type out of the netapp package
Filer type is moved from the netapp package to the main package,
because it is used for decoding the filer configuration in main
program.
